Update project from combined storage to cloud-first
