+++ cloudlink-cloudfirst-storage/src/main/java/com/gluonhq/cloudfirst/CombinedStorage.java
@@ -24,9 +24,9 @@
  * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
  * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
-package com.gluonhq.combinedstorage;
+package com.gluonhq.cloudfirst;
 
-import com.gluonhq.combinedstorage.views.AppViewManager;
+import com.gluonhq.cloudfirst.views.AppViewManager;
 import com.gluonhq.charm.glisten.application.MobileApplication;
 import com.gluonhq.charm.glisten.visual.Swatch;
 import javafx.scene.Scene;
