+++ cloudlink-cloudfirst-storage/src/main/java/com/gluonhq/cloudfirst/model/Model.java
@@ -24,43 +24,17 @@
  * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
  * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
-package com.gluonhq.combinedstorage.views;
+package com.gluonhq.cloudfirst.model;
 
-import com.gluonhq.charm.glisten.control.CharmListCell;
-import com.gluonhq.combinedstorage.model.Note;
-import java.time.format.DateTimeFormatter;
-import java.util.Locale;
-import javafx.scene.control.Label;
+import javafx.beans.property.ObjectProperty;
+import javafx.beans.property.SimpleObjectProperty;
 
-public class HeaderCell extends CharmListCell<Note> {
-
-    private final Label label;
-    private Note currentItem;
-    private final DateTimeFormatter dateFormat;
+public class Model {
     
-    public HeaderCell() {
-        label = new Label();
-        dateFormat = DateTimeFormatter.ofPattern("EEEE, MMM dd", Locale.ENGLISH);
-    }
-
-    @Override
-    public void updateItem(Note item, boolean empty) {
-        super.updateItem(item, empty);
-        currentItem = item;
-        if (!empty && item != null) {
-            updateWithSettings();
-            setGraphic(label);
-        } else {
-            setGraphic(null);
-        }
-    }
-
-    private void updateWithSettings() {
-        if (currentItem != null) {
-            label.setText(dateFormat.format(currentItem.getCreationDate()));
-        } else {
-            label.setText("");
-        }
+    private final ObjectProperty<Note> activeNote = new SimpleObjectProperty<>();
+    
+    public ObjectProperty<Note> activeNote() {
+        return activeNote;
     }
     
 }
