+++ cloudlink-cloudfirst-storage/src/main/java/com/gluonhq/cloudfirst/views/HeaderCell.java
@@ -1,5 +1,5 @@
-/*
- * Copyright (c) 2020 Gluon
+/**
+ * Copyright (c) 2017, Gluon
  * All rights reserved.
  *
  * Redistribution and use in source and binary forms, with or without
@@ -24,31 +24,43 @@
  * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
  * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
-module com.gluonhq.combinedstorage {
+package com.gluonhq.cloudfirst.views;
 
-    requires javafx.controls;
-    requires javafx.fxml;
+import com.gluonhq.charm.glisten.control.CharmListCell;
+import com.gluonhq.cloudfirst.model.Note;
+import java.time.format.DateTimeFormatter;
+import java.util.Locale;
+import javafx.scene.control.Label;
 
-    requires com.gluonhq.charm.glisten;
-    requires com.gluonhq.cloudlink.client;
-    requires com.gluonhq.connect;
-    requires com.gluonhq.glisten.afterburner;
+public class HeaderCell extends CharmListCell<Note> {
 
-    requires com.gluonhq.attach.display;
-    requires com.gluonhq.attach.lifecycle;
-    requires com.gluonhq.attach.statusbar;
-    requires com.gluonhq.attach.storage;
-    requires com.gluonhq.attach.util;
+    private final Label label;
+    private Note currentItem;
+    private final DateTimeFormatter dateFormat;
+    
+    public HeaderCell() {
+        label = new Label();
+        dateFormat = DateTimeFormatter.ofPattern("EEEE, MMM dd", Locale.ENGLISH);
+    }
 
-    requires java.json;
-    requires java.annotation;
-    requires afterburner.mfx;
+    @Override
+    public void updateItem(Note item, boolean empty) {
+        super.updateItem(item, empty);
+        currentItem = item;
+        if (!empty && item != null) {
+            updateWithSettings();
+            setGraphic(label);
+        } else {
+            setGraphic(null);
+        }
+    }
 
-    exports com.gluonhq.combinedstorage;
-    exports com.gluonhq.combinedstorage.model to afterburner.mfx;
-    exports com.gluonhq.combinedstorage.service to afterburner.mfx;
-    exports com.gluonhq.combinedstorage.views to afterburner.mfx;
-
-    opens com.gluonhq.combinedstorage.model to com.gluonhq.cloudlink.client;
-    opens com.gluonhq.combinedstorage.views to javafx.fxml, afterburner.mfx, com.gluonhq.glisten.afterburner;
+    private void updateWithSettings() {
+        if (currentItem != null) {
+            label.setText(dateFormat.format(currentItem.getCreationDate()));
+        } else {
+            label.setText("");
+        }
+    }
+    
 }
