+++ cloudlink-cloudfirst-storage/src/main/java/com/gluonhq/cloudfirst/views/NoteCell.java
@@ -24,15 +24,15 @@
  * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
  * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
-package com.gluonhq.combinedstorage.views;
+package com.gluonhq.cloudfirst.views;
 
 import com.gluonhq.attach.display.DisplayService;
 import com.gluonhq.attach.util.Services;
 import com.gluonhq.charm.glisten.control.CharmListCell;
 import com.gluonhq.charm.glisten.control.ListTile;
 import com.gluonhq.charm.glisten.visual.MaterialDesignIcon;
-import com.gluonhq.combinedstorage.model.Note;
-import com.gluonhq.combinedstorage.service.Service;
+import com.gluonhq.cloudfirst.model.Note;
+import com.gluonhq.cloudfirst.service.Service;
 import javafx.beans.value.ChangeListener;
 import javafx.geometry.Pos;
 import javafx.scene.control.Button;
