+++ cloudlink-cloudfirst-storage/src/main/java/com/gluonhq/cloudfirst/views/NotesPresenter.java
@@ -24,7 +24,7 @@
  * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
  * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
-package com.gluonhq.combinedstorage.views;
+package com.gluonhq.cloudfirst.views;
 
 import com.gluonhq.charm.glisten.afterburner.GluonPresenter;
 import com.gluonhq.charm.glisten.animation.BounceInLeftTransition;
@@ -33,10 +33,10 @@ import com.gluonhq.charm.glisten.control.CharmListView;
 import com.gluonhq.charm.glisten.control.FloatingActionButton;
 import com.gluonhq.charm.glisten.mvc.View;
 import com.gluonhq.charm.glisten.visual.MaterialDesignIcon;
-import com.gluonhq.combinedstorage.CombinedStorage;
-import com.gluonhq.combinedstorage.model.Model;
-import com.gluonhq.combinedstorage.model.Note;
-import com.gluonhq.combinedstorage.service.Service;
+import com.gluonhq.cloudfirst.CombinedStorage;
+import com.gluonhq.cloudfirst.model.Model;
+import com.gluonhq.cloudfirst.model.Note;
+import com.gluonhq.cloudfirst.service.Service;
 import javafx.fxml.FXML;
 import javafx.scene.control.Label;
