+++ cloudlink-cloudfirst-storage/src/main/resources/com/gluonhq/cloudfirst/views/edition.css
