+++ cloudlink-cloudfirst-storage/src/main/resources/com/gluonhq/cloudfirst/views/edition.fxml
@@ -38,7 +38,7 @@
 <?import javafx.scene.layout.HBox?>
 <?import javafx.scene.layout.VBox?>
 
-<View fx:id="edition" maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="600.0" prefWidth="350.0" xmlns="http://javafx.com/javafx/8" xmlns:fx="http://javafx.com/fxml/1" fx:controller="com.gluonhq.combinedstorage.views.EditionPresenter">
+<View fx:id="edition" maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="600.0" prefWidth="350.0" xmlns="http://javafx.com/javafx/8" xmlns:fx="http://javafx.com/fxml/1" fx:controller="com.gluonhq.cloudfirst.views.EditionPresenter">
    <center>
       <VBox prefHeight="200.0" spacing="10.0" BorderPane.alignment="CENTER">
          <children>
